% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{calcall_var}
\alias{calcall_var}
\title{calcall_var}
\usage{
calcall_var(inpt, var_name_v, var_val_l)
}
\arguments{
\item{inpt}{is the input formula, with the variables}

\item{var_name_v}{is the vector that contains the variables name in the order of apparition in the formula. If the variable appears multiple times in the formula, it has to be specified in this vector, see examples.}

\item{var_val_l}{is the list containing the vectors containing the values of each variable, for each point you want to calculate. The vectors has to be given in the same order has the variable in var_name_v.}
}
\description{
Does the same thing as calcall function but calculates the formula that have variables. The values of the variables have to be given in a list of vectors, see examples.
}
\examples{

print(calcall_var(inpt="(6+m*-(4-imp))+3/jp", var_name_v=c("m", "imp", "jp"), 
                  var_val_l=list(

                                 c(1:6), 

                                 c(3, 4, 2, 5, 6, 1),

                                 c(6:1))))

 [1] "5.5"  "6.6"  "0.75" "11"   "17.5" "-9"

print(calcall_var(inpt="(6+m*-(4-imp))+3/jp+jp", var_name_v=c("m", "imp", "jp", "jp"), 
                   var_val_l=list(
 
                                  c(1:6), 
 
                                  c(3, 4, 2, 5, 6, 1),
 
                                  c(6:1))))

[1] "11.5" "11.6" "4.75" "14" "19.5" "-8"

}
