% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{datf_folder}
\alias{datf_folder}
\title{datf_folder}
\usage{
datf_folder(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the input dataframe}
}
\description{
Folds a dataframe, see examples.
}
\examples{

print(datf_folder(inpt_datf = iris))

    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
1            5.1         3.5          1.4         0.2     setosa
2            6.9         3.1          4.9         1.5 versicolor
3            4.7         3.2          1.3         0.2     setosa
4            5.1         3.5          1.4         0.3     setosa
5            7.2         3.0          5.8         1.6  virginica
6            5.8         2.7          5.1         1.9  virginica
7            5.4         3.0          4.5         1.5 versicolor
8            6.7         3.1          5.6         2.4  virginica
9            6.0         3.0          4.8         1.8  virginica
10           5.4         3.4          1.5         0.4     setosa
11           6.9         3.1          5.4         2.1  virginica
12           5.8         2.7          5.1         1.9  virginica
13           6.4         3.1          5.5         1.8  virginica
14           5.7         2.6          3.5         1.0 versicolor
15           5.4         3.9          1.7         0.4     setosa
16           5.7         2.8          4.1         1.3 versicolor
17           5.1         3.7          1.5         0.4     setosa
18           4.4         3.0          1.3         0.2     setosa
19           5.7         3.8          1.7         0.3     setosa
20           5.1         3.8          1.5         0.3     setosa
21           5.4         3.4          1.7         0.2     setosa
22           6.7         3.1          4.7         1.5 versicolor
23           6.0         3.4          4.5         1.6 versicolor
24           6.9         3.1          4.9         1.5 versicolor
25           4.8         3.4          1.9         0.2     setosa
26           5.8         2.7          5.1         1.9  virginica
27           5.0         3.4          1.6         0.4     setosa
28           5.8         2.8          5.1         2.4  virginica
29           6.3         2.3          4.4         1.3 versicolor
30           4.7         3.2          1.6         0.2     setosa
31           4.8         3.0          1.4         0.3     setosa
32           5.4         3.4          1.5         0.4     setosa
33           6.1         2.6          5.6         1.4  virginica
34           6.1         3.0          4.6         1.4 versicolor
35           6.0         2.2          4.0         1.0 versicolor
36           5.0         3.2          1.2         0.2     setosa
37           5.5         3.5          1.3         0.2     setosa
38           5.8         2.8          5.1         2.4  virginica
39           6.2         3.4          5.4         2.3  virginica
40           5.1         3.4          1.5         0.2     setosa
41           5.0         3.5          1.3         0.3     setosa
42           4.5         2.3          1.3         0.3     setosa
43           4.9         3.6          1.4         0.1     setosa
44           5.0         3.5          1.6         0.6     setosa
45           5.7         3.0          4.2         1.2 versicolor
46           6.4         2.8          5.6         2.1  virginica
47           6.2         3.4          5.4         2.3  virginica
48           4.6         3.2          1.4         0.2     setosa
49           6.4         3.2          5.3         2.3  virginica
50           5.5         4.2          1.4         0.2     setosa
51           7.7         3.0          6.1         2.3  virginica
52           5.9         3.0          4.2         1.5 versicolor
53           6.5         3.0          5.5         1.8  virginica
54           5.4         3.9          1.7         0.4     setosa
55           6.5         2.8          4.6         1.5 versicolor
56           5.8         2.6          4.0         1.2 versicolor
57           5.7         2.8          4.5         1.3 versicolor
58           4.9         2.4          3.3         1.0 versicolor
59           6.7         3.1          5.6         2.4  virginica
60           6.1         3.0          4.9         1.8  virginica
61           5.8         2.8          5.1         2.4  virginica
62           5.9         3.0          4.2         1.5 versicolor
63           5.2         4.1          1.5         0.1     setosa
64           6.9         3.1          4.9         1.5 versicolor
65           5.6         2.9          3.6         1.3 versicolor
66           5.4         3.4          1.7         0.2     setosa
67           5.6         3.0          4.5         1.5 versicolor
68           5.8         2.7          4.1         1.0 versicolor
69           6.2         2.2          4.5         1.5 versicolor
70           6.2         2.2          4.5         1.5 versicolor
71           5.9         3.2          4.8         1.8 versicolor
72           6.1         2.8          4.0         1.3 versicolor
73           6.3         2.5          4.9         1.5 versicolor
74           5.0         3.0          1.6         0.2     setosa
75           4.6         3.4          1.4         0.3     setosa
76           6.4         3.2          5.3         2.3  virginica
77           6.7         3.1          4.7         1.5 versicolor
78           5.5         4.2          1.4         0.2     setosa
79           6.0         2.9          4.5         1.5 versicolor
80           5.4         3.9          1.7         0.4     setosa
81           5.5         3.5          1.3         0.2     setosa
82           6.3         3.3          6.0         2.5  virginica
83           5.8         2.7          3.9         1.2 versicolor
84           6.0         2.7          5.1         1.6 versicolor
85           6.8         2.8          4.8         1.4 versicolor
86           6.1         3.0          4.6         1.4 versicolor
87           6.7         3.1          4.7         1.5 versicolor
88           5.1         3.8          1.6         0.2     setosa
89           6.8         2.8          4.8         1.4 versicolor
90           6.9         3.2          5.7         2.3  virginica
91           6.0         3.4          4.5         1.6 versicolor
92           6.1         3.0          4.6         1.4 versicolor
93           5.8         2.6          4.0         1.2 versicolor
94           5.0         2.3          3.3         1.0 versicolor
95           5.7         3.0          4.2         1.2 versicolor
96           5.7         3.0          4.2         1.2 versicolor
97           5.7         2.9          4.2         1.3 versicolor
98           6.4         2.8          5.6         2.2  virginica
99           5.1         3.4          1.5         0.2     setosa
100          5.7         2.8          4.1         1.3 versicolor
101          6.5         2.8          4.6         1.5 versicolor
102          4.8         3.4          1.9         0.2     setosa
103          4.4         2.9          1.4         0.2     setosa
104          5.1         2.5          3.0         1.1 versicolor
105          7.4         2.8          6.1         1.9  virginica
106          7.6         3.0          6.6         2.1  virginica
107          4.9         2.5          4.5         1.7  virginica
108          7.3         2.9          6.3         1.8  virginica
109          4.8         3.4          1.9         0.2     setosa
110          5.7         4.4          1.5         0.4     setosa
111          6.5         3.2          5.1         2.0  virginica
112          6.9         3.2          5.7         2.3  virginica
113          5.9         3.2          4.8         1.8 versicolor
114          7.1         3.0          5.9         2.1  virginica
115          5.8         2.8          5.1         2.4  virginica
116          4.8         3.4          1.9         0.2     setosa
117          4.3         3.0          1.1         0.1     setosa
118          6.6         2.9          4.6         1.3 versicolor
119          5.1         2.5          3.0         1.1 versicolor
120          6.0         2.2          5.0         1.5  virginica
121          5.1         3.4          1.5         0.2     setosa
122          6.3         2.7          4.9         1.8  virginica
123          6.7         3.3          5.7         2.1  virginica
124          6.1         2.6          5.6         1.4  virginica
125          5.0         3.3          1.4         0.2     setosa
126          7.2         3.2          6.0         1.8  virginica
127          6.2         2.8          4.8         1.8  virginica
128          6.1         3.0          4.9         1.8  virginica
129          5.0         3.4          1.6         0.4     setosa
130          6.2         2.2          4.5         1.5 versicolor
131          7.4         2.8          6.1         1.9  virginica
132          6.6         2.9          4.6         1.3 versicolor
133          6.7         3.3          5.7         2.1  virginica
134          6.3         3.3          4.7         1.6 versicolor
135          5.7         2.9          4.2         1.3 versicolor
136          7.2         3.6          6.1         2.5  virginica
137          6.5         3.0          5.5         1.8  virginica
138          6.4         3.1          5.5         1.8  virginica
139          5.5         4.2          1.4         0.2     setosa
140          5.8         2.7          5.1         1.9  virginica
141          5.0         2.0          3.5         1.0 versicolor
142          6.9         3.1          5.1         2.3  virginica
143          5.8         2.7          5.1         1.9  virginica
144          6.8         3.2          5.9         2.3  virginica
145          6.7         3.3          5.7         2.5  virginica
146          5.1         3.3          1.7         0.5     setosa
147          5.1         3.8          1.9         0.4     setosa
148          6.5         3.0          5.2         2.0  virginica
149          4.6         3.6          1.0         0.2     setosa
150          5.9         3.0          5.1         1.8  virginica

}
