% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{individual_cloning}
\alias{individual_cloning}
\title{individual_cloning}
\usage{
individual_cloning(inpt_datf, col_vars = c(), label_var, hmn)
}
\arguments{
\item{inpt_datf}{is the input dataset as a dataframe}

\item{col_vars}{is a vector containing the colnames or the column numbers of the variables}

\item{label_var}{is a either the colnames or the column number of the label variable}

\item{hmn}{is how many of new individual from the same label will be generated}
}
\description{
Allow to generate individuals with the same label as those existig and having as values at variables, a value generated with a normal distribution having as parameters the mean for the variable A for the individual I and the same goes for the standard deviation, see examples.
}
\examples{

datf <- iris
datf[, 5] <- as.character(datf[, 5])
datf <- individual_cloning(inpt_datf = datf, col_vars = c(1:4), label_var = 5, hmn = 3)
print(datf)
nrow(datf)
nrow(iris)

     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
1        5.100000    3.500000     1.400000   0.2000000     setosa
2        4.900000    3.000000     1.400000   0.2000000     setosa
3        4.700000    3.200000     1.300000   0.2000000     setosa
4        4.600000    3.100000     1.500000   0.2000000     setosa
5        5.000000    3.600000     1.400000   0.2000000     setosa
6        5.400000    3.900000     1.700000   0.4000000     setosa
7        4.600000    3.400000     1.400000   0.3000000     setosa
8        5.000000    3.400000     1.500000   0.2000000     setosa
9        4.400000    2.900000     1.400000   0.2000000     setosa
10       4.900000    3.100000     1.500000   0.1000000     setosa
11       5.400000    3.700000     1.500000   0.2000000     setosa
12       4.800000    3.400000     1.600000   0.2000000     setosa
13       4.800000    3.000000     1.400000   0.1000000     setosa
14       4.300000    3.000000     1.100000   0.1000000     setosa
15       5.800000    4.000000     1.200000   0.2000000     setosa
16       5.700000    4.400000     1.500000   0.4000000     setosa
17       5.400000    3.900000     1.300000   0.4000000     setosa
18       5.100000    3.500000     1.400000   0.3000000     setosa
19       5.700000    3.800000     1.700000   0.3000000     setosa
20       5.100000    3.800000     1.500000   0.3000000     setosa
21       5.400000    3.400000     1.700000   0.2000000     setosa
22       5.100000    3.700000     1.500000   0.4000000     setosa
23       4.600000    3.600000     1.000000   0.2000000     setosa
24       5.100000    3.300000     1.700000   0.5000000     setosa
25       4.800000    3.400000     1.900000   0.2000000     setosa
26       5.000000    3.000000     1.600000   0.2000000     setosa
27       5.000000    3.400000     1.600000   0.4000000     setosa
28       5.200000    3.500000     1.500000   0.2000000     setosa
29       5.200000    3.400000     1.400000   0.2000000     setosa
30       4.700000    3.200000     1.600000   0.2000000     setosa
31       4.800000    3.100000     1.600000   0.2000000     setosa
32       5.400000    3.400000     1.500000   0.4000000     setosa
33       5.200000    4.100000     1.500000   0.1000000     setosa
34       5.500000    4.200000     1.400000   0.2000000     setosa
35       4.900000    3.100000     1.500000   0.2000000     setosa
36       5.000000    3.200000     1.200000   0.2000000     setosa
37       5.500000    3.500000     1.300000   0.2000000     setosa
38       4.900000    3.600000     1.400000   0.1000000     setosa
39       4.400000    3.000000     1.300000   0.2000000     setosa
40       5.100000    3.400000     1.500000   0.2000000     setosa
41       5.000000    3.500000     1.300000   0.3000000     setosa
42       4.500000    2.300000     1.300000   0.3000000     setosa
43       4.400000    3.200000     1.300000   0.2000000     setosa
44       5.000000    3.500000     1.600000   0.6000000     setosa
45       5.100000    3.800000     1.900000   0.4000000     setosa
46       4.800000    3.000000     1.400000   0.3000000     setosa
47       5.100000    3.800000     1.600000   0.2000000     setosa
48       4.600000    3.200000     1.400000   0.2000000     setosa
49       5.300000    3.700000     1.500000   0.2000000     setosa
50       5.000000    3.300000     1.400000   0.2000000     setosa
51       7.000000    3.200000     4.700000   1.4000000 versicolor
52       6.400000    3.200000     4.500000   1.5000000 versicolor
53       6.900000    3.100000     4.900000   1.5000000 versicolor
54       5.500000    2.300000     4.000000   1.3000000 versicolor
55       6.500000    2.800000     4.600000   1.5000000 versicolor
56       5.700000    2.800000     4.500000   1.3000000 versicolor
57       6.300000    3.300000     4.700000   1.6000000 versicolor
58       4.900000    2.400000     3.300000   1.0000000 versicolor
59       6.600000    2.900000     4.600000   1.3000000 versicolor
60       5.200000    2.700000     3.900000   1.4000000 versicolor
61       5.000000    2.000000     3.500000   1.0000000 versicolor
62       5.900000    3.000000     4.200000   1.5000000 versicolor
63       6.000000    2.200000     4.000000   1.0000000 versicolor
64       6.100000    2.900000     4.700000   1.4000000 versicolor
65       5.600000    2.900000     3.600000   1.3000000 versicolor
66       6.700000    3.100000     4.400000   1.4000000 versicolor
67       5.600000    3.000000     4.500000   1.5000000 versicolor
68       5.800000    2.700000     4.100000   1.0000000 versicolor
69       6.200000    2.200000     4.500000   1.5000000 versicolor
70       5.600000    2.500000     3.900000   1.1000000 versicolor
71       5.900000    3.200000     4.800000   1.8000000 versicolor
72       6.100000    2.800000     4.000000   1.3000000 versicolor
73       6.300000    2.500000     4.900000   1.5000000 versicolor
74       6.100000    2.800000     4.700000   1.2000000 versicolor
75       6.400000    2.900000     4.300000   1.3000000 versicolor
76       6.600000    3.000000     4.400000   1.4000000 versicolor
77       6.800000    2.800000     4.800000   1.4000000 versicolor
78       6.700000    3.000000     5.000000   1.7000000 versicolor
79       6.000000    2.900000     4.500000   1.5000000 versicolor
80       5.700000    2.600000     3.500000   1.0000000 versicolor
81       5.500000    2.400000     3.800000   1.1000000 versicolor
82       5.500000    2.400000     3.700000   1.0000000 versicolor
83       5.800000    2.700000     3.900000   1.2000000 versicolor
84       6.000000    2.700000     5.100000   1.6000000 versicolor
85       5.400000    3.000000     4.500000   1.5000000 versicolor
86       6.000000    3.400000     4.500000   1.6000000 versicolor
87       6.700000    3.100000     4.700000   1.5000000 versicolor
88       6.300000    2.300000     4.400000   1.3000000 versicolor
89       5.600000    3.000000     4.100000   1.3000000 versicolor
90       5.500000    2.500000     4.000000   1.3000000 versicolor
91       5.500000    2.600000     4.400000   1.2000000 versicolor
92       6.100000    3.000000     4.600000   1.4000000 versicolor
93       5.800000    2.600000     4.000000   1.2000000 versicolor
94       5.000000    2.300000     3.300000   1.0000000 versicolor
95       5.600000    2.700000     4.200000   1.3000000 versicolor
96       5.700000    3.000000     4.200000   1.2000000 versicolor
97       5.700000    2.900000     4.200000   1.3000000 versicolor
98       6.200000    2.900000     4.300000   1.3000000 versicolor
99       5.100000    2.500000     3.000000   1.1000000 versicolor
100      5.700000    2.800000     4.100000   1.3000000 versicolor
101      6.300000    3.300000     6.000000   2.5000000  virginica
102      5.800000    2.700000     5.100000   1.9000000  virginica
103      7.100000    3.000000     5.900000   2.1000000  virginica
104      6.300000    2.900000     5.600000   1.8000000  virginica
105      6.500000    3.000000     5.800000   2.2000000  virginica
106      7.600000    3.000000     6.600000   2.1000000  virginica
107      4.900000    2.500000     4.500000   1.7000000  virginica
108      7.300000    2.900000     6.300000   1.8000000  virginica
109      6.700000    2.500000     5.800000   1.8000000  virginica
110      7.200000    3.600000     6.100000   2.5000000  virginica
111      6.500000    3.200000     5.100000   2.0000000  virginica
112      6.400000    2.700000     5.300000   1.9000000  virginica
113      6.800000    3.000000     5.500000   2.1000000  virginica
114      5.700000    2.500000     5.000000   2.0000000  virginica
115      5.800000    2.800000     5.100000   2.4000000  virginica
116      6.400000    3.200000     5.300000   2.3000000  virginica
117      6.500000    3.000000     5.500000   1.8000000  virginica
118      7.700000    3.800000     6.700000   2.2000000  virginica
119      7.700000    2.600000     6.900000   2.3000000  virginica
120      6.000000    2.200000     5.000000   1.5000000  virginica
121      6.900000    3.200000     5.700000   2.3000000  virginica
122      5.600000    2.800000     4.900000   2.0000000  virginica
123      7.700000    2.800000     6.700000   2.0000000  virginica
124      6.300000    2.700000     4.900000   1.8000000  virginica
125      6.700000    3.300000     5.700000   2.1000000  virginica
126      7.200000    3.200000     6.000000   1.8000000  virginica
127      6.200000    2.800000     4.800000   1.8000000  virginica
128      6.100000    3.000000     4.900000   1.8000000  virginica
129      6.400000    2.800000     5.600000   2.1000000  virginica
130      7.200000    3.000000     5.800000   1.6000000  virginica
131      7.400000    2.800000     6.100000   1.9000000  virginica
132      7.900000    3.800000     6.400000   2.0000000  virginica
133      6.400000    2.800000     5.600000   2.2000000  virginica
134      6.300000    2.800000     5.100000   1.5000000  virginica
135      6.100000    2.600000     5.600000   1.4000000  virginica
136      7.700000    3.000000     6.100000   2.3000000  virginica
137      6.300000    3.400000     5.600000   2.4000000  virginica
138      6.400000    3.100000     5.500000   1.8000000  virginica
139      6.000000    3.000000     4.800000   1.8000000  virginica
140      6.900000    3.100000     5.400000   2.1000000  virginica
141      6.700000    3.100000     5.600000   2.4000000  virginica
142      6.900000    3.100000     5.100000   2.3000000  virginica
143      5.800000    2.700000     5.100000   1.9000000  virginica
144      6.800000    3.200000     5.900000   2.3000000  virginica
145      6.700000    3.300000     5.700000   2.5000000  virginica
146      6.700000    3.000000     5.200000   2.3000000  virginica
147      6.300000    2.500000     5.000000   1.9000000  virginica
148      6.500000    3.000000     5.200000   2.0000000  virginica
149      6.200000    3.400000     5.400000   2.3000000  virginica
150      5.900000    3.000000     5.100000   1.8000000  virginica
151      4.601009    3.727368     1.268078   0.3122136     setosa
210      4.613076    3.989209     1.555392   0.2953775     setosa
310      4.722235    3.602591     1.479940   0.2471369     setosa
513      5.660667    2.449398     4.241485   1.5317590 versicolor
511      5.987887    3.016099     3.690411   1.5357972 versicolor
512      5.803584    2.828602     4.024589   1.2767213 versicolor
1013     6.851160    3.287923     5.157840   1.5365199  virginica
1011     7.119751    3.460045     4.990113   1.2895762  virginica
1012     7.370573    3.140464     5.680828   1.8674812  virginica
[1] 159
[1] 150

}
