% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{individual_equalizer_max}
\alias{individual_equalizer_max}
\title{individual_equalizer_max}
\usage{
individual_equalizer_max(inpt_datf, label_var, hmn)
}
\description{
Remove the individual that are in exess according to a given value, see examples
}
\examples{

print(individual_equalizer_max(inpt_datf = datf, label_var = 5, hmn = 15))

   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
1           5.0         3.2          1.2         0.2     setosa
2           5.5         3.5          1.3         0.2     setosa
3           4.9         3.6          1.4         0.1     setosa
4           4.4         3.0          1.3         0.2     setosa
5           5.1         3.4          1.5         0.2     setosa
6           5.0         3.5          1.3         0.3     setosa
7           4.5         2.3          1.3         0.3     setosa
8           4.4         3.2          1.3         0.2     setosa
9           5.0         3.5          1.6         0.6     setosa
10          5.1         3.8          1.9         0.4     setosa
11          4.8         3.0          1.4         0.3     setosa
12          5.1         3.8          1.6         0.2     setosa
13          4.6         3.2          1.4         0.2     setosa
14          5.3         3.7          1.5         0.2     setosa
15          5.0         3.3          1.4         0.2     setosa
16          6.0         3.4          4.5         1.6 versicolor
17          6.7         3.1          4.7         1.5 versicolor
18          6.3         2.3          4.4         1.3 versicolor
19          5.6         3.0          4.1         1.3 versicolor
20          5.5         2.5          4.0         1.3 versicolor
21          5.5         2.6          4.4         1.2 versicolor
22          6.1         3.0          4.6         1.4 versicolor
23          5.8         2.6          4.0         1.2 versicolor
24          5.0         2.3          3.3         1.0 versicolor
25          5.6         2.7          4.2         1.3 versicolor
26          5.7         3.0          4.2         1.2 versicolor
27          5.7         2.9          4.2         1.3 versicolor
28          6.2         2.9          4.3         1.3 versicolor
29          5.1         2.5          3.0         1.1 versicolor
30          5.7         2.8          4.1         1.3 versicolor
31          7.7         3.0          6.1         2.3  virginica
32          6.3         3.4          5.6         2.4  virginica
33          6.4         3.1          5.5         1.8  virginica
34          6.0         3.0          4.8         1.8  virginica
35          6.9         3.1          5.4         2.1  virginica
36          6.7         3.1          5.6         2.4  virginica
37          6.9         3.1          5.1         2.3  virginica
38          5.8         2.7          5.1         1.9  virginica
39          6.8         3.2          5.9         2.3  virginica
40          6.7         3.3          5.7         2.5  virginica
41          6.7         3.0          5.2         2.3  virginica
42          6.3         2.5          5.0         1.9  virginica
43          6.5         3.0          5.2         2.0  virginica
44          6.2         3.4          5.4         2.3  virginica
45          5.9         3.0          5.1         1.8  virginica

}
