% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{individual_equalizer_min}
\alias{individual_equalizer_min}
\title{individual_equalizer_min}
\usage{
individual_equalizer_min(inpt_datf, col_vars = c(), label_var, untl)
}
\arguments{
\item{inpt_datf}{is the input dataset as a dataframe}

\item{col_vars}{is a vector containing the colnames or the column numbers of the variables}

\item{label_var}{is a either the colnames or the column number of the label variable}

\item{untl}{is how many individual from the same label the dataset has to have, at minimum}
}
\description{
Allow to increase the number of inividual from any label to a certain point based on the individual_cloning function from the same package (Rmach)
}
\examples{

datf <- iris
datf[, 5] <- as.character(datf[, 5])
datf <- individual_equalizer_min(inpt_datf = datf, col_vars = c(1:4), label_var = 5, untl = 120)
print(datf)
nrow(datf)
nrow(iris)

     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
1       5.100000    3.500000     1.400000  0.20000000     setosa
2       4.900000    3.000000     1.400000  0.20000000     setosa
3       4.700000    3.200000     1.300000  0.20000000     setosa
4       4.600000    3.100000     1.500000  0.20000000     setosa
5       5.000000    3.600000     1.400000  0.20000000     setosa
6       5.400000    3.900000     1.700000  0.40000000     setosa
7       4.600000    3.400000     1.400000  0.30000000     setosa
8       5.000000    3.400000     1.500000  0.20000000     setosa
9       4.400000    2.900000     1.400000  0.20000000     setosa
10      4.900000    3.100000     1.500000  0.10000000     setosa
11      5.400000    3.700000     1.500000  0.20000000     setosa
12      4.800000    3.400000     1.600000  0.20000000     setosa
13      4.800000    3.000000     1.400000  0.10000000     setosa
14      4.300000    3.000000     1.100000  0.10000000     setosa
15      5.800000    4.000000     1.200000  0.20000000     setosa
16      5.700000    4.400000     1.500000  0.40000000     setosa
17      5.400000    3.900000     1.300000  0.40000000     setosa
18      5.100000    3.500000     1.400000  0.30000000     setosa
19      5.700000    3.800000     1.700000  0.30000000     setosa
20      5.100000    3.800000     1.500000  0.30000000     setosa
21      5.400000    3.400000     1.700000  0.20000000     setosa
22      5.100000    3.700000     1.500000  0.40000000     setosa
23      4.600000    3.600000     1.000000  0.20000000     setosa
24      5.100000    3.300000     1.700000  0.50000000     setosa
25      4.800000    3.400000     1.900000  0.20000000     setosa
26      5.000000    3.000000     1.600000  0.20000000     setosa
27      5.000000    3.400000     1.600000  0.40000000     setosa
28      5.200000    3.500000     1.500000  0.20000000     setosa
29      5.200000    3.400000     1.400000  0.20000000     setosa
30      4.700000    3.200000     1.600000  0.20000000     setosa
31      4.800000    3.100000     1.600000  0.20000000     setosa
32      5.400000    3.400000     1.500000  0.40000000     setosa
33      5.200000    4.100000     1.500000  0.10000000     setosa
34      5.500000    4.200000     1.400000  0.20000000     setosa
35      4.900000    3.100000     1.500000  0.20000000     setosa
36      5.000000    3.200000     1.200000  0.20000000     setosa
37      5.500000    3.500000     1.300000  0.20000000     setosa
38      4.900000    3.600000     1.400000  0.10000000     setosa
39      4.400000    3.000000     1.300000  0.20000000     setosa
40      5.100000    3.400000     1.500000  0.20000000     setosa
41      5.000000    3.500000     1.300000  0.30000000     setosa
42      4.500000    2.300000     1.300000  0.30000000     setosa
43      4.400000    3.200000     1.300000  0.20000000     setosa
44      5.000000    3.500000     1.600000  0.60000000     setosa
45      5.100000    3.800000     1.900000  0.40000000     setosa
46      4.800000    3.000000     1.400000  0.30000000     setosa
47      5.100000    3.800000     1.600000  0.20000000     setosa
48      4.600000    3.200000     1.400000  0.20000000     setosa
49      5.300000    3.700000     1.500000  0.20000000     setosa
50      5.000000    3.300000     1.400000  0.20000000     setosa
51      7.000000    3.200000     4.700000  1.40000000 versicolor
52      6.400000    3.200000     4.500000  1.50000000 versicolor
53      6.900000    3.100000     4.900000  1.50000000 versicolor
54      5.500000    2.300000     4.000000  1.30000000 versicolor
55      6.500000    2.800000     4.600000  1.50000000 versicolor
56      5.700000    2.800000     4.500000  1.30000000 versicolor
57      6.300000    3.300000     4.700000  1.60000000 versicolor
58      4.900000    2.400000     3.300000  1.00000000 versicolor
59      6.600000    2.900000     4.600000  1.30000000 versicolor
60      5.200000    2.700000     3.900000  1.40000000 versicolor
61      5.000000    2.000000     3.500000  1.00000000 versicolor
62      5.900000    3.000000     4.200000  1.50000000 versicolor
63      6.000000    2.200000     4.000000  1.00000000 versicolor
64      6.100000    2.900000     4.700000  1.40000000 versicolor
65      5.600000    2.900000     3.600000  1.30000000 versicolor
66      6.700000    3.100000     4.400000  1.40000000 versicolor
67      5.600000    3.000000     4.500000  1.50000000 versicolor
68      5.800000    2.700000     4.100000  1.00000000 versicolor
69      6.200000    2.200000     4.500000  1.50000000 versicolor
70      5.600000    2.500000     3.900000  1.10000000 versicolor
71      5.900000    3.200000     4.800000  1.80000000 versicolor
72      6.100000    2.800000     4.000000  1.30000000 versicolor
73      6.300000    2.500000     4.900000  1.50000000 versicolor
74      6.100000    2.800000     4.700000  1.20000000 versicolor
75      6.400000    2.900000     4.300000  1.30000000 versicolor
76      6.600000    3.000000     4.400000  1.40000000 versicolor
77      6.800000    2.800000     4.800000  1.40000000 versicolor
78      6.700000    3.000000     5.000000  1.70000000 versicolor
79      6.000000    2.900000     4.500000  1.50000000 versicolor
80      5.700000    2.600000     3.500000  1.00000000 versicolor
81      5.500000    2.400000     3.800000  1.10000000 versicolor
82      5.500000    2.400000     3.700000  1.00000000 versicolor
83      5.800000    2.700000     3.900000  1.20000000 versicolor
84      6.000000    2.700000     5.100000  1.60000000 versicolor
85      5.400000    3.000000     4.500000  1.50000000 versicolor
86      6.000000    3.400000     4.500000  1.60000000 versicolor
87      6.700000    3.100000     4.700000  1.50000000 versicolor
88      6.300000    2.300000     4.400000  1.30000000 versicolor
89      5.600000    3.000000     4.100000  1.30000000 versicolor
90      5.500000    2.500000     4.000000  1.30000000 versicolor
91      5.500000    2.600000     4.400000  1.20000000 versicolor
92      6.100000    3.000000     4.600000  1.40000000 versicolor
93      5.800000    2.600000     4.000000  1.20000000 versicolor
94      5.000000    2.300000     3.300000  1.00000000 versicolor
95      5.600000    2.700000     4.200000  1.30000000 versicolor
96      5.700000    3.000000     4.200000  1.20000000 versicolor
97      5.700000    2.900000     4.200000  1.30000000 versicolor
98      6.200000    2.900000     4.300000  1.30000000 versicolor
99      5.100000    2.500000     3.000000  1.10000000 versicolor
100     5.700000    2.800000     4.100000  1.30000000 versicolor
101     6.300000    3.300000     6.000000  2.50000000  virginica
102     5.800000    2.700000     5.100000  1.90000000  virginica
103     7.100000    3.000000     5.900000  2.10000000  virginica
104     6.300000    2.900000     5.600000  1.80000000  virginica
105     6.500000    3.000000     5.800000  2.20000000  virginica
106     7.600000    3.000000     6.600000  2.10000000  virginica
107     4.900000    2.500000     4.500000  1.70000000  virginica
108     7.300000    2.900000     6.300000  1.80000000  virginica
109     6.700000    2.500000     5.800000  1.80000000  virginica
110     7.200000    3.600000     6.100000  2.50000000  virginica
111     6.500000    3.200000     5.100000  2.00000000  virginica
112     6.400000    2.700000     5.300000  1.90000000  virginica
113     6.800000    3.000000     5.500000  2.10000000  virginica
114     5.700000    2.500000     5.000000  2.00000000  virginica
115     5.800000    2.800000     5.100000  2.40000000  virginica
116     6.400000    3.200000     5.300000  2.30000000  virginica
117     6.500000    3.000000     5.500000  1.80000000  virginica
118     7.700000    3.800000     6.700000  2.20000000  virginica
119     7.700000    2.600000     6.900000  2.30000000  virginica
120     6.000000    2.200000     5.000000  1.50000000  virginica
121     6.900000    3.200000     5.700000  2.30000000  virginica
122     5.600000    2.800000     4.900000  2.00000000  virginica
123     7.700000    2.800000     6.700000  2.00000000  virginica
124     6.300000    2.700000     4.900000  1.80000000  virginica
125     6.700000    3.300000     5.700000  2.10000000  virginica
126     7.200000    3.200000     6.000000  1.80000000  virginica
127     6.200000    2.800000     4.800000  1.80000000  virginica
128     6.100000    3.000000     4.900000  1.80000000  virginica
129     6.400000    2.800000     5.600000  2.10000000  virginica
130     7.200000    3.000000     5.800000  1.60000000  virginica
131     7.400000    2.800000     6.100000  1.90000000  virginica
132     7.900000    3.800000     6.400000  2.00000000  virginica
133     6.400000    2.800000     5.600000  2.20000000  virginica
134     6.300000    2.800000     5.100000  1.50000000  virginica
135     6.100000    2.600000     5.600000  1.40000000  virginica
136     7.700000    3.000000     6.100000  2.30000000  virginica
137     6.300000    3.400000     5.600000  2.40000000  virginica
138     6.400000    3.100000     5.500000  1.80000000  virginica
139     6.000000    3.000000     4.800000  1.80000000  virginica
140     6.900000    3.100000     5.400000  2.10000000  virginica
141     6.700000    3.100000     5.600000  2.40000000  virginica
142     6.900000    3.100000     5.100000  2.30000000  virginica
143     5.800000    2.700000     5.100000  1.90000000  virginica
144     6.800000    3.200000     5.900000  2.30000000  virginica
145     6.700000    3.300000     5.700000  2.50000000  virginica
146     6.700000    3.000000     5.200000  2.30000000  virginica
147     6.300000    2.500000     5.000000  1.90000000  virginica
148     6.500000    3.000000     5.200000  2.00000000  virginica
149     6.200000    3.400000     5.400000  2.30000000  virginica
150     5.900000    3.000000     5.100000  1.80000000  virginica
151     5.119546    3.240896     1.659373  0.25516050     setosa
152     4.902088    4.003746     1.228617  0.35778383     setosa
153     4.834331    3.698540     1.547812  0.33339113     setosa
154     5.134884    3.180819     1.588032  0.18761885     setosa
155     5.488401    3.298369     1.683031  0.18180736     setosa
156     4.758992    3.086108     1.434159  0.25348240     setosa
157     4.817610    3.052438     1.470246  0.18414810     setosa
158     5.372952    3.815612     1.344489  0.12705451     setosa
159     5.203751    3.331928     1.384586  0.26145797     setosa
160     5.154693    4.326639     1.585445  0.10767788     setosa
161     4.651867    2.915629     1.333128  0.24085761     setosa
162     4.703818    3.295307     1.524695  0.53200346     setosa
163     5.299254    3.127387     1.436154  0.32571756     setosa
164     4.576459    3.690579     1.500380  0.24860844     setosa
165     4.821700    3.891746     1.277726  0.34434218     setosa
166     5.195495    2.693142     1.518095  0.11628275     setosa
167     4.751171    4.076332     1.437831  0.29611751     setosa
168     4.895746    3.340168     1.505157  0.32204518     setosa
169     5.084452    2.649230     1.253577  0.34230634     setosa
170     4.994526    3.283612     1.466568  0.10785695     setosa
171     4.914249    3.713116     1.456736  0.13825711     setosa
172     5.168494    3.384539     1.391309  0.36352904     setosa
173     4.868237    3.608825     1.580430  0.16346689     setosa
174     4.922010    3.812630     1.385674  0.17966376     setosa
175     4.782539    3.520596     1.166369  0.19443475     setosa
176     4.999012    2.953373     1.276890  0.04813659     setosa
177     4.237476    3.501651     1.603897 -0.02137016     setosa
178     4.161835    2.900175     1.340508  0.31471652     setosa
179     5.326641    2.690628     1.367918  0.30229792     setosa
180     5.144879    2.889594     1.627228  0.29699450     setosa
181     5.032020    3.092995     1.262743  0.13014888     setosa
182     4.912576    4.102884     1.592814  0.46510333     setosa
183     4.886276    3.643501     1.362697  0.45850332     setosa
184     5.067843    3.644076     1.284018  0.11802271     setosa
185     4.870130    3.261045     1.387769  0.24945158     setosa
186     4.203276    3.532647     1.759381  0.22793382     setosa
187     5.147728    2.949748     1.344759  0.14613345     setosa
188     5.044451    3.821792     1.690910  0.27432788     setosa
189     5.144534    3.260319     1.486522  0.15193060     setosa
190     4.749463    3.242690     1.558031  0.29964703     setosa
191     5.012355    4.056773     1.568806  0.28175520     setosa
192     5.286178    3.657418     1.556329  0.25865612     setosa
193     4.739473    3.599081     1.361732  0.11096506     setosa
194     4.763743    3.719912     1.532282  0.23680057     setosa
195     4.352927    3.606171     1.443575  0.22201153     setosa
196     5.420318    3.234039     1.257110  0.29332868     setosa
197     5.032471    4.002458     1.149330  0.14118440     setosa
198     4.679526    3.634655     1.503754  0.19732104     setosa
199     4.655581    2.890624     1.538909  0.10855489     setosa
200     5.432263    3.587195     1.448039  0.15201721     setosa
201     5.030955    3.620666     1.379309  0.22296525     setosa
202     5.117052    3.640415     1.680914  0.22426164     setosa
203     4.206403    3.577511     1.579905  0.34627623     setosa
204     5.345245    3.207691     1.351151  0.10816533     setosa
205     5.287934    3.630390     1.494184  0.31610331     setosa
206     4.371540    3.674677     1.483436  0.12756906     setosa
207     4.458787    3.512193     1.499114  0.35598241     setosa
208     4.694526    4.189214     1.065203  0.32728599     setosa
209     5.199256    3.164026     1.523074 -0.00277085     setosa
210     4.857067    3.279462     1.431379  0.28051926     setosa
211     5.120333    3.079011     1.256199  0.26650341     setosa
212     5.526492    3.715932     1.385397  0.10935802     setosa
213     4.255062    3.442076     1.032584  0.22553491     setosa
214     5.547997    3.899931     1.805604  0.14245435     setosa
215     5.056086    3.556886     1.485842  0.25052054     setosa
216     4.602273    3.582194     1.637627  0.10750785     setosa
217     5.707143    3.272366     1.495331  0.24957136     setosa
218     4.529437    3.295707     1.370119  0.22733484     setosa
219     4.815724    3.274761     1.264803  0.19839835     setosa
220     5.219331    3.678528     1.534777  0.31111961     setosa
221     7.469393    3.164987     4.045576  1.29571858 versicolor
222     6.435686    2.759148     4.152976  1.38660991 versicolor
223     6.004909    2.617229     3.374965  1.50931533 versicolor
224     5.998960    2.889328     4.787927  0.99816956 versicolor
225     6.066878    2.738260     4.317450  1.35360632 versicolor
226     6.558577    3.004756     3.518091  1.10350572 versicolor
227     5.226591    2.937582     4.211646  1.82617395 versicolor
228     6.519901    3.085536     4.666132  1.47417398 versicolor
229     6.212108    2.297953     3.256134  1.57999643 versicolor
230     6.234065    2.904038     3.899946  1.61728632 versicolor
231     5.891353    2.871663     3.585063  1.15322879 versicolor
232     5.495659    2.332178     4.373762  1.40539853 versicolor
233     5.484945    3.186158     5.022759  1.03428734 versicolor
234     5.002003    2.716631     4.221475  1.13953629 versicolor
235     6.289043    3.017459     3.910062  1.38286708 versicolor
236     5.700736    3.131150     4.960207  1.14958223 versicolor
237     5.500216    3.190272     4.253273  1.18245190 versicolor
238     6.445503    2.960724     4.621510  1.21795268 versicolor
239     5.889688    2.752965     4.360846  1.21917725 versicolor
240     5.217994    2.727503     4.018054  1.19655177 versicolor
241     5.628761    2.782079     4.503714  1.21105694 versicolor
242     5.922639    2.647391     3.774616  1.48842237 versicolor
243     6.021925    2.565549     3.937052  1.45849084 versicolor
244     6.330301    2.687627     4.026615  0.94678432 versicolor
245     6.304311    2.635169     3.998727  1.45603553 versicolor
246     6.663896    3.297885     3.907486  1.16979322 versicolor
247     5.376404    2.885587     3.866554  1.05112744 versicolor
248     4.695327    2.578715     3.943357  1.16919180 versicolor
249     6.278448    3.381682     3.893139  1.31728551 versicolor
250     5.808922    2.342279     4.329488  1.36901786 versicolor
251     6.257850    3.299147     4.763327  1.45358673 versicolor
252     5.397398    2.181731     5.237967  1.63885805 versicolor
253     6.318406    3.370869     4.403785  1.71528585 versicolor
254     6.030213    2.934996     5.690094  1.18095022 versicolor
255     6.322254    2.643724     4.712019  1.30067547 versicolor
256     5.483814    3.540120     3.935919  1.36104088 versicolor
257     4.923149    2.834738     3.978205  1.09514320 versicolor
258     5.102353    3.275399     4.167623  1.69802624 versicolor
259     6.503755    2.772905     4.500401  1.10261134 versicolor
260     6.024940    2.379938     3.663719  1.24096925 versicolor
261     6.155505    2.960939     4.628437  1.63876689 versicolor
262     6.547596    2.753326     3.814345  1.50055748 versicolor
263     7.340028    3.049036     4.128880  1.43704378 versicolor
264     6.771703    2.744679     3.755760  1.35657812 versicolor
265     6.526113    3.315310     4.723554  1.13676188 versicolor
266     5.737681    2.732723     4.619607  1.20118401 versicolor
267     5.118896    3.053538     5.153921  1.24286955 versicolor
268     6.557536    2.506483     3.775426  1.25665234 versicolor
269     6.773637    3.056770     3.907444  1.48359009 versicolor
270     5.231083    2.716242     3.701491  1.43445828 versicolor
271     6.373044    2.810367     3.823155  1.48776176 versicolor
272     6.689764    2.329003     4.315204  1.20003129 versicolor
273     5.909787    2.877026     3.921463  1.44035219 versicolor
274     5.985060    3.408963     4.312826  1.14822888 versicolor
275     5.720711    3.047025     4.502301  1.30692891 versicolor
276     6.075586    2.625810     3.462166  1.13883320 versicolor
277     5.979742    3.037604     4.337108  1.17174718 versicolor
278     5.944742    3.187138     4.131605  1.40617115 versicolor
279     5.377366    2.850410     4.848731  1.31109047 versicolor
280     5.911520    2.601061     3.978657  1.19677413 versicolor
281     6.299276    3.083130     3.767828  1.21669672 versicolor
282     6.508117    2.717810     4.400327  1.15816277 versicolor
283     5.564065    2.991926     3.244794  0.97614826 versicolor
284     5.636803    3.041730     3.675623  1.52144698 versicolor
285     6.249670    2.545928     4.021866  1.48874150 versicolor
286     5.779178    3.126088     4.456842  1.35907598 versicolor
287     5.056560    3.158496     4.029340  1.09487926 versicolor
288     6.256082    2.754099     3.546839  1.10515518 versicolor
289     6.727157    3.127967     4.478930  1.36983039 versicolor
290     6.644075    2.156546     4.073352  1.24130902 versicolor
291     6.086009    2.661626     6.272420  1.43328200  virginica
292     6.415624    3.507285     4.970803  2.29244152  virginica
293     7.783730    3.194127     6.263952  2.12710505  virginica
294     6.714708    2.207256     4.695838  1.57280728  virginica
295     6.892027    3.146945     5.963832  2.03720894  virginica
296     6.384602    2.842640     5.424208  1.34455702  virginica
297     7.151880    2.761441     5.193842  2.65759524  virginica
298     7.000909    3.538284     5.949645  2.37981867  virginica
299     6.267784    3.471146     5.832588  1.97858577  virginica
300     6.684294    3.095409     5.918461  1.79584906  virginica
301     6.653542    3.193293     5.478747  2.02974253  virginica
302     6.932936    2.532998     5.398907  2.58686242  virginica
303     6.171339    3.401070     5.778270  2.14575174  virginica
304     6.321461    3.238482     5.728325  1.77370288  virginica
305     6.939597    3.105226     5.153168  2.30218152  virginica
306     4.983468    2.869016     5.249331  2.33602954  virginica
307     7.057275    3.000195     5.368063  2.29811745  virginica
308     5.648449    3.022504     4.670324  2.44199827  virginica
309     7.023223    3.038748     6.549980  1.74164740  virginica
310     6.621430    2.928325     4.114293  1.65060008  virginica
311     5.947210    2.572431     6.035025  1.67473550  virginica
312     6.720834    2.791217     4.373968  1.80139289  virginica
313     7.277691    3.013233     6.057093  2.41664038  virginica
314     6.036578    3.034487     5.680667  2.14347484  virginica
315     7.523033    2.906421     5.746571  2.19174990  virginica
316     6.148008    3.219150     5.385260  2.29487465  virginica
317     6.653134    3.286357     5.439343  2.01415643  virginica
318     7.665406    2.418833     4.912548  2.04701493  virginica
319     6.962181    3.122207     5.926113  2.14427668  virginica
320     6.968055    3.394053     5.176526  2.28774948  virginica
321     8.433217    3.190685     6.154875  1.86645175  virginica
322     5.865485    3.206422     6.182362  2.06380350  virginica
323     6.357587    3.105502     6.086674  2.22194560  virginica
324     7.000027    3.093890     5.694556  1.95490517  virginica
325     5.329756    3.313431     7.114499  1.82374316  virginica
326     7.063835    2.978432     6.702789  1.97846514  virginica
327     6.643032    3.331938     5.319034  1.98032475  virginica
328     5.812732    2.605752     4.698275  2.04751518  virginica
329     5.922603    2.951062     4.789723  1.86828922  virginica
330     6.534338    3.077621     4.735738  1.96590508  virginica
331     6.566409    2.869386     5.256565  2.30887779  virginica
332     5.873025    2.576689     5.399706  1.51365277  virginica
333     6.436762    2.807203     5.237271  1.70436243  virginica
334     6.700115    2.741499     6.361120  2.57743789  virginica
335     6.800498    2.964161     6.726096  2.01077453  virginica
336     6.817689    3.044292     5.651350  1.64623491  virginica
337     6.589657    2.978472     6.011304  2.51979646  virginica
338     8.263734    3.121411     5.285361  1.93618630  virginica
339     7.027356    2.891612     5.821978  1.92039311  virginica
340     4.943241    2.503378     5.732430  1.80385345  virginica
341     7.071175    2.628713     6.012994  2.06170238  virginica
342     6.074115    3.436504     5.791817  1.23968953  virginica
343     6.853310    2.681229     5.643604  1.21275207  virginica
344     6.254123    3.365158     5.832863  2.67274454  virginica
345     6.511558    2.738037     5.355683  1.85846301  virginica
346     5.842295    3.300082     4.540820  2.12329402  virginica
347     6.423004    3.294433     6.394560  1.76478497  virginica
348     5.833874    3.222916     5.861218  1.69319220  virginica
349     6.478021    3.028388     6.606609  2.06623919  virginica
350     7.784342    2.902471     5.142493  1.91602616  virginica
351     6.775815    3.445127     5.519265  2.13719655  virginica
352     7.014933    2.715428     6.798085  2.04147119  virginica
353     7.689606    2.506295     5.531764  1.88075834  virginica
354     7.506985    2.788839     5.837837  2.47057469  virginica
355     7.242421    2.782457     6.390016  1.66938074  virginica
356     6.400116    2.353697     4.388649  2.24717026  virginica
357     7.384851    3.077118     5.716925  2.36297064  virginica
358     6.892294    3.466955     4.959172  2.13813060  virginica
359     5.904443    3.286340     4.911794  1.90991134  virginica
360     6.292600    2.938076     5.710938  2.61396630  virginica
[1] 360
[1] 150

}
