% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{Rmach_det}
\alias{Rmach_det}
\title{Rmach_det}
\usage{
Rmach_det(inpt_matr)
}
\arguments{
\item{inpt_matr}{}
}
\description{
Calculates the determinant of any square matrix, see examples
}
\examples{

mtr_test2 <- matrix(nrow = 2, ncol = 2, data = c(4:6, 37))
mtr_test3 <- matrix(nrow = 3, ncol = 3, data = c(4:6, 37, 12, 33, 1, 2, 3))
mtr_test4 <- matrix(nrow = 4, ncol = 4, data = c(4:6, 37, 12, 33, 1, 2, 3, 8, 7, 8, 7, 11, 22, 21))
mtr_test5 <- matrix(nrow = 5, ncol = 5, data = c(1:25))
mtr_test6 <- matrix(nrow = 6, ncol = 6, data = c(1:36))
mtr_test7 <- matrix(nrow = 7, ncol = 7, data = c(1:49))
mtr_test8 <- matrix(nrow = 8, ncol = 8, data = c(1:64))
mtr_test9 <- matrix(nrow = 9, ncol = 9, data = c(1:81))

det(mtr_test2)
[1] 118
print(Rmach_det(inpt_matr = mtr_test2))
[1] 118

det(mtr_test3)
[1] -138 
print(Rmach_det(inpt_matr = mtr_test3))
[1] -138

det(mtr_test4)
[1] -20001
print(Rmach_det(inpt_matr = mtr_test4))
[1] -20001

det(mtr_test5)
[1] 0
print(Rmach_det(inpt_matr = mtr_test5))
[1] 0

det(mtr_test6)
[1] 0
print(Rmach_det(inpt_matr = mtr_test6))
[1] 0

det(mtr_test7)
[1] 0
print(Rmach_det(inpt_matr = mtr_test7))
[1] 0

det(mtr_test8)
[1] 0
print(Rmach_det(inpt_matr = mtr_test8))
[1] 0

det(mtr_test9)
[1] 0
print(Rmach_det(inpt_matr = mtr_test2))
[1] 0

}
