% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\docType{class}
\name{sample_Rmach-class}
\alias{sample_Rmach-class}
\title{v_Rmach_fold}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{train_prop}{is the training proportion}

\item{n_fold}{is the number of distinc pair of training and test dataset that will be outputed}
}
\description{
Allow to create uniform sampling dataset for cross validation,
train and test, see examples and variables
}
