% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\docType{class}
\name{sample_Rmach-class}
\alias{sample_Rmach-class}
\title{v_Rmach_fold}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{train_prop}{is the training proportion}

\item{n_fold}{is the number of distinc pair of training and test dataset that will be outputed}
}
\description{
Allow to create uniform sampling dataset for cross validation,
train and test, see examples and variables
}
\examples{

lst_test <- v_Rmach_fold(inpt_datf = iris[1:25,],
             train_prop = 0.7,
             n_fold = 4)

print(lst_test)

$sample1
An object of class "sample_Rmach"
Slot "train":
     Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
24            5.1         3.3          1.7         0.5  setosa           0
18            5.1         3.5          1.4         0.3  setosa           0
12            4.8         3.4          1.6         0.2  setosa           0
19            5.7         3.8          1.7         0.3  setosa           0
20            5.1         3.8          1.5         0.3  setosa           0
5             5.0         3.6          1.4         0.2  setosa           0
4             4.6         3.1          1.5         0.2  setosa           0
23            4.6         3.6          1.0         0.2  setosa           0
18.1          5.1         3.5          1.4         0.3  setosa           0
1             5.1         3.5          1.4         0.2  setosa           0
7             4.6         3.4          1.4         0.3  setosa           0
14            4.3         3.0          1.1         0.1  setosa           0
7.1           4.6         3.4          1.4         0.3  setosa           0
4.1           4.6         3.1          1.5         0.2  setosa           0
19.1          5.7         3.8          1.7         0.3  setosa           0
9             4.4         2.9          1.4         0.2  setosa           0
8             5.0         3.4          1.5         0.2  setosa           0
16            5.7         4.4          1.5         0.4  setosa           0

Slot "test":
   Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
7           4.6         3.4          1.4         0.3  setosa           1
12          4.8         3.4          1.6         0.2  setosa           1
8           5.0         3.4          1.5         0.2  setosa           1
14          4.3         3.0          1.1         0.1  setosa           1
11          5.4         3.7          1.5         0.2  setosa           1
25          4.8         3.4          1.9         0.2  setosa           1
23          4.6         3.6          1.0         0.2  setosa           1

Slot "train_ids":
 [1] 24 18 12 19 20  5  4 23 18  1  7 14  7  4 19  9  8 16

Slot "test_ids":
[1]  7 12  8 14 11 25 23


$sample2
An object of class "sample_Rmach"
Slot "train":
     Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
20            5.1         3.8          1.5         0.3  setosa           0
8             5.0         3.4          1.5         0.2  setosa           0
2             4.9         3.0          1.4         0.2  setosa           0
11            5.4         3.7          1.5         0.2  setosa           0
22            5.1         3.7          1.5         0.4  setosa           0
13            4.8         3.0          1.4         0.1  setosa           0
24            5.1         3.3          1.7         0.5  setosa           0
2.1           4.9         3.0          1.4         0.2  setosa           0
7             4.6         3.4          1.4         0.3  setosa           0
2.2           4.9         3.0          1.4         0.2  setosa           0
22.1          5.1         3.7          1.5         0.4  setosa           0
22.2          5.1         3.7          1.5         0.4  setosa           0
24.1          5.1         3.3          1.7         0.5  setosa           0
22.3          5.1         3.7          1.5         0.4  setosa           0
3             4.7         3.2          1.3         0.2  setosa           0
3.1           4.7         3.2          1.3         0.2  setosa           0
11.1          5.4         3.7          1.5         0.2  setosa           0
6             5.4         3.9          1.7         0.4  setosa           0

Slot "test":
   Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
8           5.0         3.4          1.5         0.2  setosa           1
12          4.8         3.4          1.6         0.2  setosa           1
1           5.1         3.5          1.4         0.2  setosa           1
11          5.4         3.7          1.5         0.2  setosa           1
2           4.9         3.0          1.4         0.2  setosa           1
18          5.1         3.5          1.4         0.3  setosa           1
20          5.1         3.8          1.5         0.3  setosa           1

Slot "train_ids":
 [1] 20  8  2 11 22 13 24  2  7  2 22 22 24 22  3  3 11  6

Slot "test_ids":
[1]  8 12  1 11  2 18 20


$sample3
An object of class "sample_Rmach"
Slot "train":
     Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
5             5.0         3.6          1.4         0.2  setosa           0
14            4.3         3.0          1.1         0.1  setosa           0
16            5.7         4.4          1.5         0.4  setosa           0
4             4.6         3.1          1.5         0.2  setosa           0
16.1          5.7         4.4          1.5         0.4  setosa           0
15            5.8         4.0          1.2         0.2  setosa           0
3             4.7         3.2          1.3         0.2  setosa           0
18            5.1         3.5          1.4         0.3  setosa           0
25            4.8         3.4          1.9         0.2  setosa           0
23            4.6         3.6          1.0         0.2  setosa           0
4.1           4.6         3.1          1.5         0.2  setosa           0
24            5.1         3.3          1.7         0.5  setosa           0
20            5.1         3.8          1.5         0.3  setosa           0
7             4.6         3.4          1.4         0.3  setosa           0
19            5.7         3.8          1.7         0.3  setosa           0
21            5.4         3.4          1.7         0.2  setosa           0
23.1          4.6         3.6          1.0         0.2  setosa           0
11            5.4         3.7          1.5         0.2  setosa           0

Slot "test":
   Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
18          5.1         3.5          1.4         0.3  setosa           1
21          5.4         3.4          1.7         0.2  setosa           1
5           5.0         3.6          1.4         0.2  setosa           1
12          4.8         3.4          1.6         0.2  setosa           1
14          4.3         3.0          1.1         0.1  setosa           1
2           4.9         3.0          1.4         0.2  setosa           1
8           5.0         3.4          1.5         0.2  setosa           1

Slot "train_ids":
 [1]  5 14 16  4 16 15  3 18 25 23  4 24 20  7 19 21 23 11

Slot "test_ids":
[1] 18 21  5 12 14  2  8


$sample4
An object of class "sample_Rmach"
Slot "train":
     Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
18            5.1         3.5          1.4         0.3  setosa           0
18.1          5.1         3.5          1.4         0.3  setosa           0
13            4.8         3.0          1.4         0.1  setosa           0
7             4.6         3.4          1.4         0.3  setosa           0
18.2          5.1         3.5          1.4         0.3  setosa           0
2             4.9         3.0          1.4         0.2  setosa           0
19            5.7         3.8          1.7         0.3  setosa           0
9             4.4         2.9          1.4         0.2  setosa           0
23            4.6         3.6          1.0         0.2  setosa           0
15            5.8         4.0          1.2         0.2  setosa           0
16            5.7         4.4          1.5         0.4  setosa           0
15.1          5.8         4.0          1.2         0.2  setosa           0
8             5.0         3.4          1.5         0.2  setosa           0
9.1           4.4         2.9          1.4         0.2  setosa           0
10            4.9         3.1          1.5         0.1  setosa           0
14            4.3         3.0          1.1         0.1  setosa           0
11            5.4         3.7          1.5         0.2  setosa           0
12            4.8         3.4          1.6         0.2  setosa           0

Slot "test":
   Sepal.Length Sepal.Width Petal.Length Petal.Width Species test_status
9           4.4         2.9          1.4         0.2  setosa           1
13          4.8         3.0          1.4         0.1  setosa           1
4           4.6         3.1          1.5         0.2  setosa           1
19          5.7         3.8          1.7         0.3  setosa           1
22          5.1         3.7          1.5         0.4  setosa           1
11          5.4         3.7          1.5         0.2  setosa           1
5           5.0         3.6          1.4         0.2  setosa           1

Slot "train_ids":
 [1] 18 18 13  7 18  2 19  9 23 15 16 15  8  9 10 14 11 12

Slot "test_ids":
[1]  9 13  4 19 22 11  5

}
